Add retention schedule ordering to its own helper
